Separate the Custom Response class to include in controllers and improve intelephense
