+++ src/App/App.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 require __DIR__ . '/../../vendor/autoload.php';
 require __DIR__ . '/DotEnv.php';
+require_once __DIR__.'/CustomResponse.php';
 require_once __DIR__.'/ResponseFactory.php';
 $app = require __DIR__ . '/Container.php';
 $customErrorHandler = require __DIR__ . '/ErrorHandler.php';
