+++ src/Controller/Home.php
@@ -4,10 +4,9 @@ declare(strict_types=1);
 
 namespace App\Controller;
 
-use App\Helper\JsonResponse;
 use Pimple\Psr11\Container;
-use Psr\Http\Message\ResponseInterface as Response;
 use Psr\Http\Message\ServerRequestInterface as Request;
+use App\CustomResponse as Response;
 
 final class Home
 {
